# [jdbc.row] Reuse (:cols builder) during fetch-all-columns!