Another easy way to get -15% execution time during selects. The boost comes from looking up `(:cols builder)` once per row, not per each cell. As an extra bonus, counting the column vector directly is slighgly faster than protocol-based `(next.jdbc.rs/column-count builder)`.

10k rows benchmark against current master:

```
Before (master):
Time per call: 4.35 ms   Alloc per call: 8,446,823b   Iterations: 2315
Time per call: 4.33 ms   Alloc per call: 8,446,793b   Iterations: 2371
Time per call: 4.32 ms   Alloc per call: 8,446,796b   Iterations: 2321

After (this PR):
Time per call: 3.67 ms   Alloc per call: 8,446,794b   Iterations: 2761
Time per call: 3.69 ms   Alloc per call: 8,446,795b   Iterations: 2745
Time per call: 3.68 ms   Alloc per call: 8,446,793b   Iterations: 2731
```